+++ R/import-testthat.R
@@ -1,5 +1,5 @@
 #' @import testthat
-#' @importFrom rlang quo enquo expr enexpr eval_tidy list2 has_length :=
+#' @importFrom rlang quo enquo enquos expr enexpr eval_tidy list2 has_length :=
 NULL
 
 #' @importFrom methods findMethod getClasses getClass extends