+++ R/spec-driver-class.R
@@ -1,7 +1,0 @@
-spec_driver_class <- list(
-  inherits_from_driver = function(ctx) {
-    expect_s4_class(ctx$drv, "DBIDriver")
-  },
-
-  NULL
-)