+++ R/spec-driver.R
@@ -1,6 +1,5 @@
 #' @format NULL
 spec_driver <- c(
-  spec_driver_class,
   spec_driver_constructor,
   spec_driver_data_type,
   spec_driver_get_info,